requirements.txt
pyautogui
keyboard
pynput
cryptography
python-xlib